requirements.txt
selenium==4.9.1
pillow==9.0.1
pytesseract==0.3.9
